requirements/requirements_load.txt
numpy==1.23.3
pandas==1.5.0
ping3==4.0.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41